requirements.txt
Flask==3.0.3
grpcio==1.67.1
grpcio-tools==1.67.1
requests==2.32.3